NO SQL/fast api + JWT auth + beanie + monogodb/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
bcrypt==4.1.2
beanie==1.25.0
click==8.1.7
dnspython==2.5.0
ecdsa==0.18.0
exceptiongroup==1.2.0
fastapi==0.109.2
h11==0.14.0
idna==3.6
lazy-model==0.2.0
motor==3.3.2
passlib==1.7.4
pyasn1==0.5.1
pydantic==2.6.1
pydantic_core==2.16.2
PyJWT==2.8.0
pymongo==4.6.1
python-jose==3.3.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.36.3
toml==0.10.2
typing_extensions==4.9.0
uvicorn==0.27.0.post1
